rename parser/imports to parser/fastscan
